Implement xarray <-> dataset
This new convention supports bi-directional conversion between xarrays and datasets, under the assumption that an xarray is an image.

Currently, this does flip the axes order for C-style (default) indexed xarrays, as Java uses F-style indexing.  This behavior conforms to the current status for regular numpy arrays.

Right now the conversion also assumes that the axes are linear, such that the axes can be defined with just an origin and a scale (aX + b).

Any non-numeric axes labels are currently lost (e.g., if you have coords of [R, G, B] they become [0, 1, 2] upon conversion)